dotfile generator: fish: add checks for the number of arguments to the setup script to prevent error messages from appearing

+++ gen/setup.fish
@@ -1,13 +1,13 @@
 fish setup-functions.fish
 fish setup-abbreviations.fish
-if test $argv[1] = generate-git
-	if test $argv[2] = aliases
+if test (count $argv) -ge 1; and test $argv[1] = generate-git
+	if test (count $argv) -ge 2; and test $argv[2] = aliases
 		fish setup-git-aliases.fish
 	else
 		fish setup-git-abbreviations.fish
 	end
-else if test $argv[1] = copy-git
-	if test $argv[2] = aliases; and test -f ../.config/fish/conf.d/git-aliases.fish
+else if test (count $argv) -ge 1; and test $argv[1] = copy-git
+	if test (count $argv) -ge 2; and test $argv[2] = aliases; and test -f ../.config/fish/conf.d/git-aliases.fish
 		cp ../.config/fish/conf.d/git-aliases.fish ~/.config/fish/conf.d/
 	else if test -f ../.config/fish/conf.d/git-abbreviations.fish
 		cp ../.config/fish/conf.d/git-abbreviations.fish ~/.config/fish/conf.d/
